Fix sign comparison warning

+++ core/src/Cuda/Kokkos_Cuda_Parallel_Team.hpp
@@ -582,7 +582,7 @@ class ParallelFor<FunctorType, Kokkos::TeamPolicy<Properties...>,
           "Kokkos::Impl::ParallelFor< Cuda > insufficient shared memory"));
     }
 
-    if (m_scratch_size[1] > m_policy.scratch_size_max(1)) {
+    if (m_scratch_size[1] > static_cast<size_t>(m_policy.scratch_size_max(1))) {
       Kokkos::Impl::throw_runtime_exception(
           std::string("Kokkos::Impl::ParallelFor< Cuda > insufficient level 1 "
                       "scratch memory"));
@@ -978,7 +978,7 @@ class ParallelReduce<CombinedFunctorReducerType,
                       "L0 scratch memory"));
     }
 
-    if (m_scratch_size[1] > m_policy.scratch_size_max(1)) {
+    if (m_scratch_size[1] > static_cast<size_t>(m_policy.scratch_size_max(1))) {
       Kokkos::Impl::throw_runtime_exception(
           std::string("Kokkos::Impl::ParallelFor< Cuda > requested too much L1 "
                       "scratch memory"));
