+++ core/src/HIP/Kokkos_HIP_ParallelReduce_Team.hpp
@@ -395,7 +395,7 @@ class ParallelReduce<CombinedFunctorReducerType,
                       "L0 scratch memory"));
     }
 
-    if (m_scratch_size[1] > m_policy.scratch_size_max(1)) {
+    if (m_scratch_size[1] > static_cast<size_t>(m_policy.scratch_size_max(1))) {
       Kokkos::Impl::throw_runtime_exception(
           std::string("Kokkos::Impl::ParallelFor< HIP > requested too much L1 "
                       "scratch memory"));
